Mock works with any EventEmitter.
It would have been hard to test programs that use Olio because they
startup and want to build an Olio handing it the `process` object. With
Arguable I have an object that is a proxy to `process` so you can test
actions by simulating events. I want ot build an Olio object with that
EventEmitter, but then have it work with mocked sockets instead of real
sockets. So, we're passing in a mock factory as an event through the
EventEmitter.