Make `--socket` an explicit argument of `listen`.
Make the `--socket` an explicit argument of the `listen` command. The
positional argument is a left over from when we where going to launch
services by having them connect to the UNIX domain socket with
instructions on how to start a child. Now children are specified as
command line arguments instead, launched immediately and the socket is
known to all.