Upgrade Arguable to 8.0.x.
Upgrade Arguable to 8.0.x. The new arguable provides a destructible for
the user to use to control the shutdown of their program. You're able to
specify the scram using a command line variable of your choosing, that
is the application developer's choosing. We use `scram` as the parameter
for Olio.

The new Arguabl also provides a facility for mocking parent/child pipes,
so that I no longer have to find a way to create a file descriptor to
test the invocation of `new Socket` for each program that uses child
sockets. Now that everything I write is Olio based, that doesn't matter
quite so much, but it does make our tests simpler, to my eyes at least.

The Olio Child process now uses Arguable to use these facilities, the
Destructible and pipe mocking, instead of being it's own beastie.