Rename `children` to `constituents`.
The name `children` confuses documentation when I'm writing about
aspects of Olio specific to UNIX child process and sensing that the
reading might not differentiate these thoughts from discussion of the
more abstract concept of an Olio constituent (née child.)

The first change I'm going to make is to rename this in the definition
file that our dear user will create to configure their Olio application.
I'll follow up with renames of classes and variables.