Disconnect after `Destrucible` completes.
Apparently, if you use `disconnect` you will not run down the entire
Node.js event loop or else I got some other aspect of `disconnect`
wrong, but it is the case that if you `Server.close` you will
immediately be removed from the round-robin algorithm so we can signal
our constituents to shutdown and so long as they shutdown their shared
servers, we can disconnect as a final act. This gives `Destructible`
time to run and time to scram.