Use new Procession async serialization.
Using the new Procession async Reader/Writer pair and placing a handler
on the sockets `"error"` event. Procession serialization used to try to
capture all errors and handle them for you, but it didn't ever manage to
do that and it turned into an encapsulated, configurated lump.

The new Reader/Writer pair is more verbose, but pretty easy to read and
use. Not sure why I'm using `raise` when I have the error handler,
though. Really when the socket is done, it's done. I don't care why.
Truncation is an error if it doesn't come at program exit time.