Do not raise reader/writer errors.
We're already logging them and we don't really care about them
otherwise. Truncation before the end of the program is a problem in
itself, the specific cause doesn't matter. Errors at the end of the
program are no problem. It means a sibling crashed. The siblings crash
will be recorded in a big way. Its effect on us doesn't need to be.